pyproject.toml
[project]
name = "my-daily-tasks-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "genai-prices>=0.0.47",
    "pydantic-ai>=1.27.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
